feat: Added new furigana parsers
